chore: rename initializers package to repositories
This change aligns with common naming conventions in the layer
architecture, where we typically have controller, service, and
repository layers.
