+++ migrate/db_generater.go
@@ -1,12 +1,12 @@
 package migrate
 
 import (
-	"github.com/docentre/docentre/initializers"
 	"github.com/docentre/docentre/models"
+	"github.com/docentre/docentre/repositories"
 )
 
 // CreateDatabase creates the tables used in this application.
 func CreateDatabase() {
-	initializers.DB.AutoMigrate(&models.User{})
-	initializers.DB.AutoMigrate(&models.Document{})
+	repositories.DB.AutoMigrate(&models.User{})
+	repositories.DB.AutoMigrate(&models.Document{})
 }
