+++ repositories/env_var.go
@@ -1,4 +1,4 @@
-package initializers
+package repositories
 
 import (
 	"log"
