+++ services/document_service.go
@@ -3,13 +3,13 @@ package services
 import (
 	"time"
 
-	"github.com/docentre/docentre/initializers"
 	"github.com/docentre/docentre/models"
+	"github.com/docentre/docentre/repositories"
 )
 
 func CreateDocument(authorID uint) (models.Document, error) {
 	doc := models.Document{AuthorID: authorID, LastEditDate: time.Now()}
-	result := initializers.DB.Create(&doc)
+	result := repositories.DB.Create(&doc)
 	if result.Error != nil {
 		return models.Document{}, result.Error
 	}
