+++ services/user_service.go
@@ -3,14 +3,14 @@ package services
 import (
 	"log"
 
-	"github.com/docentre/docentre/initializers"
 	"github.com/docentre/docentre/models"
+	"github.com/docentre/docentre/repositories"
 )
 
 func UserCreate(name string, email string, password string) (models.User, error) {
 	user := models.User{Username: name, Email: email, Password: password, Identity: "user"}
 
-	result := initializers.DB.Create(&user)
+	result := repositories.DB.Create(&user)
 
 	if result.Error != nil {
 		return models.User{}, result.Error
@@ -21,7 +21,7 @@ func UserCreate(name string, email string, password string) (models.User, error)
 
 func UserLogin(name string, password string) (models.User, error) {
 	var user models.User
-	result := initializers.DB.Where(&models.User{Username: name, Password: password}).First(&user)
+	result := repositories.DB.Where(&models.User{Username: name, Password: password}).First(&user)
 
 	if result.Error != nil {
 		return models.User{}, result.Error
@@ -33,7 +33,7 @@ func UserLogin(name string, password string) (models.User, error) {
 func GetUsersByUsername(name string) ([]models.User, error) {
 	var users []models.User
 	// perform like query
-	result := initializers.DB.Where("username LIKE ?", "%"+name+"%").Find(&users)
+	result := repositories.DB.Where("username LIKE ?", "%"+name+"%").Find(&users)
 
 	log.Println(users)
 	if result.Error != nil {
